requirements.txt
albucore==0.0.12
albumentations==1.4.10
annotated-types==0.7.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
efficientnet_pytorch==0.7.1
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
munch==4.0.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pretrainedmodels==0.7.4
pydantic==2.8.0
pydantic_core==2.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.14.0
segmentation-models-pytorch==0.3.3
six==1.16.0
sympy==1.12.1
tbb==2021.13.0
threadpoolctl==3.5.0
tifffile==2024.6.18
timm==0.9.2
tomli==2.0.1
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
